+++ src/main/java/com/coniverse/dangjang/domain/version/controller/ProdIntroController.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang.domain.intro.controller;
+package com.coniverse.dangjang.domain.version.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.coniverse.dangjang.domain.intro.dto.IntroResponse;
-import com.coniverse.dangjang.domain.intro.service.IntroService;
+import com.coniverse.dangjang.domain.version.dto.IntroResponse;
+import com.coniverse.dangjang.domain.version.service.IntroService;
 import com.coniverse.dangjang.global.dto.SuccessSingleResponse;
 
 import lombok.RequiredArgsConstructor;
