+++ src/test/java/com/coniverse/dangjang/support/ControllerTest.java
@@ -25,8 +25,6 @@ import com.coniverse.dangjang.domain.healthmetric.service.HealthConnectService;
 import com.coniverse.dangjang.domain.healthmetric.service.HealthMetricChartSearchService;
 import com.coniverse.dangjang.domain.healthmetric.service.HealthMetricRegisterService;
 import com.coniverse.dangjang.domain.healthmetric.service.HealthMetricSearchService;
-import com.coniverse.dangjang.domain.intro.controller.IntroController;
-import com.coniverse.dangjang.domain.intro.service.IntroService;
 import com.coniverse.dangjang.domain.log.controller.AppLogController;
 import com.coniverse.dangjang.domain.log.service.LogService;
 import com.coniverse.dangjang.domain.notification.controller.NotificationController;
@@ -39,6 +37,8 @@ import com.coniverse.dangjang.domain.user.repository.UserRepository;
 import com.coniverse.dangjang.domain.user.service.MypageService;
 import com.coniverse.dangjang.domain.user.service.UserSignupService;
 import com.coniverse.dangjang.domain.user.service.UserWithdrawalService;
+import com.coniverse.dangjang.domain.version.controller.IntroController;
+import com.coniverse.dangjang.domain.version.service.IntroService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
